There are multiple changes in this commit.
The user can specify which budget and which account to use: previously this was hard-coded to the default budget and the 'Natwest' account.This introduces a breaking change: the user must now supply both of these.

Add support for configuration file: require that the user pass in a YAML configuration file containing source banks and the target budgets/accounts in YNAB. At the moment we still only support one bank (natwest) uploading to one account. This is a breaking change for the arguments of the CLI.

The email address for YNAB is moved out of the secret and into the configuration file.