Remove transaction_store from the YNAB class.
This means we can construct a TransactionStore per-bank, and then keep it around to use for later operations. Previously the transaction_store was a member of YNAB and was implicitly cleared during push, which wasn't at all clear to a reader of the code. This refactor makes everything more functional and understandable.